code/requirements.txt
pandas
python-dotenv
gql
requests
requests_toolbelt